+++ robofriends/src/components/SearchBox.js
